move gcal endpoints to platform
