+++ apps/api/v2/src/modules/apps/apps.module.ts
@@ -1,5 +1,4 @@
 import { AppsRepository } from "@/modules/apps/apps.repository";
-import { GoogleCalendarOAuthController } from "@/modules/apps/controllers/gcal-oauth/gcal-oauth.controller";
 import { CredentialsRepository } from "@/modules/credentials/credentials.repository";
 import { PrismaModule } from "@/modules/prisma/prisma.module";
 import { SelectedCalendarsRepository } from "@/modules/selected-calendars/selected-calendars.repository";
@@ -10,7 +9,6 @@ import { ConfigService } from "@nestjs/config";
 @Module({
   imports: [PrismaModule, TokensModule],
   providers: [AppsRepository, ConfigService, CredentialsRepository, SelectedCalendarsRepository],
-  controllers: [GoogleCalendarOAuthController],
   exports: [],
 })
 export class AppsModule {}
